l10/2 - auth checks

+++ movies.php
@@ -3,7 +3,12 @@ $title = 'Movies';
 require 'includes/header.php';
 ?>
         <h1>Movies</h1>
-        <a href="movie-details.php">Add a New Movie</a>
+        <?php
+        // we don't need to call session_start() first because we already called it in the header above
+        if (!empty($_SESSION['username'])) {
+            echo '<a href="movie-details.php">Add a New Movie</a>';
+        } 
+        ?>
         <table class="table table-striped table-hover">
             <thead>
                 <tr>
@@ -11,7 +16,12 @@ require 'includes/header.php';
                     <th>Rating</th>
                     <th>Release Year</th>
                     <th>Genre</th>
-                    <th>Actions</th>
+                    <?php
+                    // we don't need to call session_start() first because we already called it in the header above
+                    if (!empty($_SESSION['username'])) {
+                        echo '<th>Actions</th>';
+                    } 
+                    ?>                   
                 </tr>
            </thead>                
            <tbody>
@@ -27,21 +37,28 @@ require 'includes/header.php';
                     // loop through the records, new row for each record, new column for each value
                     foreach ($movies as $movie) {
                         echo '<tr>
-                            <td>
-                                <a href="movie-details.php?movieId=' . $movie['movieId'] . '">' .                            
-                                $movie['title'] . '</a>
-                            </td>
-                            <td>' . $movie['rating'] . '</td>
-                            <td>' . $movie['releaseYear'] . '</td>
-                            <td>' . $movie['name'] . '</td>
-                            <td>
+                            <td>';
+                            if (!empty($_SESSION['username'])) {
+                                echo '<a href="movie-details.php?movieId=' . $movie['movieId'] . '">' .                            
+                                    $movie['title'] . '</a>';
+                            }
+                            else {
+                                echo $movie['title'];
+                            }
+                            echo '</td>
+                                <td>' . $movie['rating'] . '</td>
+                                <td>' . $movie['releaseYear'] . '</td>
+                                <td>' . $movie['name'] . '</td>';
+                        if (!empty($_SESSION['username'])) {                        
+                            echo '<td>
                                 <a class="btn btn-danger" 
                                     onclick="return confirmDelete()"
                                     href="delete-movie.php?movieId=' . $movie['movieId'] . '">
                                     Delete
                                 </a>
-                            </td>
-                            </tr>';
+                            </td>';
+                        }
+                        echo '</tr>';
                     }
                     $db = null;
                 }
